Fix: Frontend: Adapted authenticatio logic with refresh token

+++ backend/src/main/java/com/fitgeek/IATestPreparator/dtos/ResultResponseDto.java
@@ -3,5 +3,5 @@ package com.fitgeek.IATestPreparator.dtos;
 public record ResultResponseDto (
         int correctCount,
         int size,
-        double rate) {
+        Long rate) {
 }
